+++ frontend/src/pages/Main/Main.tsx
@@ -1,3 +1,4 @@
+import { Suspense } from 'react';
 import styled from 'styled-components';
 import Footer from '../../components/Footer';
 import Search from '../../components/search/Search';
@@ -23,7 +24,9 @@ const Main = () => {
           <div>{SUBTITLE}</div>
         </ContentContainer>
         <ErrorBoundary fallback={RankingErrorFallback}>
-          <KeywordRanking />
+          <Suspense fallback={<></>}>
+            <KeywordRanking />
+          </Suspense>
         </ErrorBoundary>
         <Search />
       </MainContainer>
