+++ frontend/src/pages/PaperDetail/PaperDetail.tsx
@@ -2,7 +2,6 @@ import { useQuery } from 'react-query';
 import { useNavigate, useSearchParams } from 'react-router-dom';
 import styled from 'styled-components';
 import Api from '../../api/api';
-import MoonLoader from '../../components/MoonLoader';
 import PreviousButtonIcon from '../../icons/PreviousButtonIcon';
 import LogoIcon from '../../icons/LogoIcon';
 import { IPaper } from '../SearchList/SearchList';
@@ -30,7 +29,7 @@ const PaperDatail = () => {
   const navigate = useNavigate();
   const [searchParams] = useSearchParams();
   const doi = searchParams.get('doi') || '';
-  const { data, isLoading } = useQuery<IPaperDetail>(
+  const { data } = useQuery<IPaperDetail>(
     ['paperDetail', doi],
     () => api.getPaperDetail({ doi }).then((res) => res.data),
     { enabled: !!doi.length },
@@ -46,23 +45,15 @@ const PaperDatail = () => {
 
   return (
     <Container>
-      {isLoading ? (
-        <MoonWrapper>
-          <MoonLoader />
-        </MoonWrapper>
-      ) : (
-        <>
-          <Header>
-            <IconButton icon={<PreviousButtonIcon />} onClick={handlePreviousButtonClick} />
-            <IconButton icon={<LogoIcon height="30" width="30" />} onClick={handleLogoClick} />
-          </Header>
-          {data && (
-            <Main>
-              <PaperInfo data={data} />
-              <ReferenceGragh />
-            </Main>
-          )}
-        </>
+      <Header>
+        <IconButton icon={<PreviousButtonIcon />} onClick={handlePreviousButtonClick} />
+        <IconButton icon={<LogoIcon height="30" width="30" />} onClick={handleLogoClick} />
+      </Header>
+      {data && (
+        <Main>
+          <PaperInfo data={data} />
+          <ReferenceGragh />
+        </Main>
       )}
     </Container>
   );
@@ -72,13 +63,6 @@ const Container = styled.div`
   display: flex;
   height: 100%;
   background-color: ${({ theme }) => theme.COLOR.primary4};
-`;
-
-const MoonWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  flex: 1;
 `;
 
 const Header = styled.header`
