+++ frontend/src/pages/SearchList/components/SearchResults.tsx
@@ -3,7 +3,6 @@ import { useQuery } from 'react-query';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 import Api, { IGetSearch } from '../../../api/api';
-import MoonLoader from '../../../components/MoonLoader';
 import Pagination from '../../../components/Pagination';
 import { createDetailQuery } from '../../../utils/createQuery';
 import { IPageInfo, IPaper } from '../SearchList';
@@ -24,48 +23,29 @@ const api = new Api();
 const SearchResults = ({ params, changePage }: SearchResultsProps) => {
   const keyword = params.keyword || '';
   const page = Number(params.page);
-  const { data, isLoading } = useQuery<IPapersData>(
-    ['papers', params],
-    () => api.getSearch(params).then((res) => res.data),
-    {
-      enabled: !isEmpty(params),
-    },
-  );
+  const { data } = useQuery<IPapersData>(['papers', params], () => api.getSearch(params).then((res) => res.data), {
+    enabled: !isEmpty(params),
+  });
 
-  return (
+  return data && data.papers.length > 0 ? (
     <>
-      {isLoading ? (
-        <MoonWrapper>
-          <MoonLoader />
-        </MoonWrapper>
-      ) : data && data.papers.length > 0 ? (
-        <>
-          <H1>Articles ({data.pageInfo.totalItems.toLocaleString() || 0})</H1>
-          <Hr />
-          <Section>
-            <Papers>
-              {data.papers.map((paper) => (
-                <Link key={paper.doi} to={createDetailQuery(paper.doi)}>
-                  <Paper data={paper} keyword={keyword} />
-                </Link>
-              ))}
-            </Papers>
-            <Pagination activePage={page} onChange={changePage} totalPages={data.pageInfo.totalPages} range={10} />
-          </Section>
-        </>
-      ) : (
-        <NoResult>&apos;{keyword}&apos;에 대한 검색 결과가 없습니다.</NoResult>
-      )}
+      <H1>Articles ({data.pageInfo.totalItems.toLocaleString() || 0})</H1>
+      <Hr />
+      <Section>
+        <Papers>
+          {data.papers.map((paper) => (
+            <Link key={paper.doi} to={createDetailQuery(paper.doi)}>
+              <Paper data={paper} keyword={keyword} />
+            </Link>
+          ))}
+        </Papers>
+        <Pagination activePage={page} onChange={changePage} totalPages={data.pageInfo.totalPages} range={10} />
+      </Section>
     </>
+  ) : (
+    <NoResult>&apos;{keyword}&apos;에 대한 검색 결과가 없습니다.</NoResult>
   );
 };
-
-const MoonWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  flex: 1;
-`;
 
 const H1 = styled.h1`
   color: ${({ theme }) => theme.COLOR.gray4};
